% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robinson_SLearner.R
\name{Robinson_SLearner}
\alias{Robinson_SLearner}
\title{Robinson estimation with Super Learner}
\usage{
Robinson_SLearner(
  X,
  D,
  Y,
  SL.DML.library = c("SL.ranger", "SL.mean"),
  DML.V = 2,
  SL.V = 2,
  core = 1,
  process = TRUE,
  weights = NULL
)
}
\arguments{
\item{X}{Control as data frame}

\item{D}{Treatment vector}

\item{Y}{Outcome vector}

\item{SL.DML.library}{SL.library to estimate nuisance functions. Use first method if all SL coefficient are zero.}

\item{DML.V}{Number of folds}

\item{SL.V}{Number of folds in each SuperLearner}

\item{weights}{Sample weights}
}
\value{
A list containing follows:
\item{data}{ Data frame including estimated nuisance}
\item{parameter_estimation}{ Estimation result}
}
\description{
\code{Robinson_SLearner} is used to implement Robinson estimation with Super Learner
}
\examples{

# Example data
X <- matrix(rnorm(200,0,1),100)
D <- 2*X[,1] + runif(100,-10,10)
Y <- 3*X[,1] + rnorm(100,0,1)

# Estimation
fit <- Robinson_SLearner(X = X, D = D, Y = Y)
fit$parameter_estimation

}
\references{
Robinson, P. M. (1988). Root-N-consistent semiparametric regression. Econometrica, 931-954.

Chernozhukov, V., Chetverikov, D., Demirer, M., Duflo, E., Hansen, C., Newey, W., & Robins, J. (2018). Double/debiased machine learning for treatment and structural parameters.

van der Laan, M. J., Polley, E. C. and Hubbard, A. E. (2007) Super Learner. Statistical Applications of Genetics and Molecular Biology, 6, article 25.
}
