% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SenstivityAIPW.R
\name{SenstivityAIPW}
\alias{SenstivityAIPW}
\title{Sensitivity Analysis for AIPW estimators}
\usage{
SenstivityAIPW(Y, M, G, Alpha, RY2, RD2)
}
\arguments{
\item{Y}{Outcome vector}

\item{M}{Plugin score vector}

\item{G}{Outcome predictors}

\item{Alpha}{Reiz representer}

\item{RY2}{R2 between prediction error of Y and unobservables}

\item{RD2}{R2 between estimated and ideal Reiz representer}
}
\value{
A list containing follows:
\item{Nuisance}{ estimated nuisance}
\item{Score}{ Score}
\item{Main}{ Estimation results}
}
\description{
\code{SenstivityAIPW} is used to implement Robinson estimation with Super Learner
Y, M, G, Alpha, RY2, RD2
}
\examples{

# Example data
X <- matrix(rnorm(200,0,1),100)
D <- 2*X[,1] + runif(100,-10,10)
Y <- 3*X[,1] + rnorm(100,0,1)

# Estimation
fit <- Robinson_SLearner(X = X, D = D, Y = Y)
fit$parameter_estimation

}
\references{
Cinelli, C., & Hazlett, C. (2020). Making sense of sensitivity: Extending omitted variable bias. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 82(1), 39-67.

Chernozhukov, V., Cinelli, C., Newey, W., Sharma, A., & Syrgkanis, V. (2021). Omitted Variable Bias in Machine Learned Causal Models. arXiv preprint arXiv:2112.13398.
}
