% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SenstivityAIPW.R
\name{SenstivityAIPW}
\alias{SenstivityAIPW}
\title{Sensitivity Analysis for AIPW estimators}
\usage{
SenstivityAIPW(Y, M, G, Alpha, RY2, RD2)
}
\arguments{
\item{Y}{Outcome vector}

\item{M}{Plugin score vector}

\item{G}{Outcome predictors}

\item{Alpha}{Reiz representer}

\item{RY2}{R2 between prediction error of Y and unobservables}

\item{RD2}{R2 between estimated and ideal Reiz representer}
}
\value{
A list containing follows:
\item{Nuisance}{ estimated nuisance}
\item{Score}{ Score}
\item{Main}{ Estimation results}
}
\description{
\code{SenstivityAIPW} is used to implement Robinson estimation with Super Learner
Y, M, G, Alpha, RY2, RD2
}
\examples{

# Example data
N <- 1000
X <- runif(N,0,1)
D <- sample(0:1,N,replace = TRUE)
Y <- D + X + rnorm(N,0,1)

Fit <- AIPW$new(Y = Y,
A = D,
W = X,
Q.SL.library = c("SL.lm"),
g.SL.library = c("SL.lm")
)
Fit$fir()$Fit$summary()

EstG <- Fit$obs_est$mu
EstM <- Fit$obs_est$mu1 - Fit$obs_est$mu0
HatD <- Fit$obs_est$p_score
EstAlpha <- D/HatD - (1-D)/(1-HatD)
FitTest <- EstPartialLinear(Y = Y, M = EstM, G = EstG, Alpha = EstAlpha, RY2 = 0.05, RD2 = 0.95)
tibble(PointATE = FitTest$Main$Theta,
PointUpper95th = FitTest$Main$Theta + 1.96*FitTest$Main$ThetaSD,
PointLower95th = FitTest$Main$Theta - 1.96*FitTest$Main$ThetaSD,
BoundUpper = FitTest$Main$Upper,
BoundLower = FitTest$Main$Lower,
BoundUpper95th = FitTest$Main$Upper + 1.96*FitTest$Main$UpperSD,
BoundLower95th = FitTest$Main$Lower - 1.96*FitTest$Main$LowerSD)


}
\references{
Cinelli, C., & Hazlett, C. (2020). Making sense of sensitivity: Extending omitted variable bias. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 82(1), 39-67.

Chernozhukov, V., Cinelli, C., Newey, W., Sharma, A., & Syrgkanis, V. (2021). Omitted Variable Bias in Machine Learned Causal Models. arXiv preprint arXiv:2112.13398.
}
